service/googleauth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyjwt==2.8.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0